# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Traveloop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    patch:
      operationId: updateCurrentUser
      tags: [Auth]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserBody'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUser'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      operationId: deleteCurrentUser
      tags: [Auth]
      summary: Delete the current user's account and all data
      responses:
        '204':
          description: Account deleted
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /users/me/photo:
    post:
      operationId: uploadProfilePhoto
      tags: [Auth]
      summary: Upload a profile photo via Cloudinary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadPhotoBody'
      responses:
        '200':
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadPhotoResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...